[#96] Entity 정리
