+++ poporazzi/Domain/Entity/Album.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 
+/// 앨범
 struct Album {
     
     /// 고유 아이디
@@ -51,28 +52,6 @@ struct Album {
             mediaFetchOption: .all,
             mediaFilterOption: .init()
         )
-    }
-}
-
-struct MediaFilterOption {
-    
-    /// 직접 촬영한 사진 포함 여부
-    var isContainSelfShooting: Bool
-    
-    /// 다운로드 사진 포함 여부
-    var isContainDownload: Bool
-    
-    /// 스크린샷 포함 여부
-    var isContainScreenshot: Bool
-    
-    init(
-        isContainSelfShooting: Bool = true,
-        isContainDownload: Bool = false,
-        isContainScreenshot: Bool = false
-    ) {
-        self.isContainSelfShooting = isContainSelfShooting
-        self.isContainDownload = isContainDownload
-        self.isContainScreenshot = isContainScreenshot
     }
 }
